Fix: tried to write the same file twice
Fix the problem that if two .proto files belong to same package, protoc
would complain: "tried to write the same file twice". This is because in
protoc-gen-md we are looping and generate markdown output file through
each request.proto_file, but the file name is determined by the package.

To fix this, first we save the content to write in md_files dictionary
with key be the output file paths. If multiple request.proto_file belong
to same package (same output file), the content would be concatnated.
After looping through all request.proto_file, we add to the response one
by one.

To make the concatnated markdown content not so weird, the title is
changed from "API Reference" to each .proto filename. So "API Reference"
would repeat multiple times.

To verify the result, say you have foo.proto as following:

    syntax = "proto2";
    import "bar.proto";
    package foo;
    message Foo {
        optional Bar bar = 1;
    }

And bar.proto as following:

    syntax = "proto2";
    package foo;
    message Bar {
        optional int32 value = 1;
    }

Run the protoc command as following to verify:

    $ protoc --md_out=. foo.proto